+++ src/components/validator_staking_desktop/hooks.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { useState } from 'react';
 
-export const validatorStakingHook = () => {
+// eslint-disable-next-line max-len
+export const validatorStakingHook = (count?: number, onChangePage?: any) => {
   const [page, setPage] = useState(0);
   const [rowsPerPage, setRowsPerPage] = useState(5);
 
@@ -19,11 +20,35 @@ export const validatorStakingHook = () => {
     setPage(0);
   };
 
+  const handleFirstPageButtonClick = (event: React.MouseEvent<HTMLButtonElement>) => {
+    onChangePage(event, 0);
+  };
+
+  const handleBackButtonClick = (event: React.MouseEvent<HTMLButtonElement>) => {
+    const prevPage = page - 1;
+    onChangePage(event, prevPage);
+    setPage(prevPage);
+  };
+
+  const handleNextButtonClick = (event: React.MouseEvent<HTMLButtonElement>) => {
+    const nextPage = page + 1;
+    onChangePage(event, nextPage);
+    setPage(nextPage);
+  };
+
+  const handleLastPageButtonClick = (event: React.MouseEvent<HTMLButtonElement>) => {
+    onChangePage(event, Math.max(0, Math.ceil(count / rowsPerPage) - 1));
+  };
+
   return {
     page,
     rowsPerPage,
     handleChangePage,
     handleChangeRowsPerPage,
+    handleFirstPageButtonClick,
+    handleBackButtonClick,
+    handleNextButtonClick,
+    handleLastPageButtonClick,
 
   };
 };
